+++ Pandora/Paradigm/Inventory.hs
@@ -3,6 +3,7 @@ module Pandora.Paradigm.Inventory (module Exports) where
 import Pandora.Paradigm.Inventory.Optics as Exports
 import Pandora.Paradigm.Inventory.Store as Exports
 import Pandora.Paradigm.Inventory.State as Exports
+import Pandora.Paradigm.Inventory.Imprint as Exports
 import Pandora.Paradigm.Inventory.Equipment as Exports
 import Pandora.Paradigm.Inventory.Environment as Exports
 import Pandora.Paradigm.Inventory.Accumulator as Exports